add new customer filters

+++ framework/Source/GPUImageVibranceFilter.h
@@ -1,9 +1,9 @@
 //
 //  GPUImageVibranceFilter.h
+//  GPUImage
 //
-//
-//  Created by github.com/r3mus on 8/14/15.
-//
+//  Created by Xin Zeng on 12/27/16.
+//  Copyright © 2016 Brad Larson. All rights reserved.
 //
 
 #import "GPUImageFilter.h"
@@ -13,8 +13,6 @@
     GLint vibranceUniform;
 }
 
-// Modifies the saturation of desaturated colors, leaving saturated colors unmodified.
-// Value -1 to 1 (-1 is minimum vibrance, 0 is no change, and 1 is maximum vibrance)
-@property (readwrite, nonatomic) GLfloat vibrance;
+@property(readwrite, nonatomic) CGFloat vibrance;
 
 @end
